+++ changelog.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased Changes]
-None currently
+
+### Updated
+
+- Sodium compat (v0.8.12-alpha.4)
 
 ## [6.0.7] - 2025-06-18